+++ x64barebones/Kernel/syscalls/sys_destroy_sem.c
@@ -1,7 +1,7 @@
 #include <semaphore.h>
 #include <syscalls.h>
 
-unsigned int sys_destroy_sem(unsigned int sem_id)
+uint64_t sys_destroy_sem(unsigned int sem_id)
 {
 	destroy_sem(sem_id);
 	return 1;
