+++ x64barebones/Kernel/syscalls/sys_get_pid.c
@@ -1,7 +1,7 @@
 #include <multitasking.h>
 #include <syscalls.h>
 
-unsigned int sys_get_pid()
+uint64_t sys_get_pid()
 {
 	return get_current_pid();
 }
