+++ x64barebones/Kernel/syscalls/sys_kill_process.c
@@ -1,6 +1,6 @@
 #include <multitasking.h>
 #include <syscalls.h>
 
-unsigned int sys_kill_process(unsigned int pid) {
+uint64_t sys_kill_process(unsigned int pid) {
 	return removeTask(pid);
 }
