+++ x64barebones/Kernel/syscalls/sys_nice.c
@@ -1,7 +1,7 @@
 #include <multitasking.h>
 #include <syscalls.h>
 
-unsigned int sys_nice(uint8_t pid, int delta)
+uint64_t sys_nice(uint8_t pid, int delta)
 {
 	return change_priority(pid, delta);
 }
