+++ x64barebones/Kernel/syscalls/sys_read_from_screen.c
@@ -1,6 +1,6 @@
 #include <multitasking.h>
 #include <syscalls.h>
 
-unsigned int sys_read_from_screen(char *buf, unsigned int count) {
+uint64_t sys_read_from_screen(char *buf, unsigned int count) {
 	return sys_read(getCurrentScreen(),buf,count);			// no requiere de la especificacion de a que pantalla escribir 
 }
